Working parser
A single source of truth
